requirements.txt
yt-dlp
ffmpeg-python
openai-whisper
anthropic
python-dotenv
tqdm
torch
flask
